refactor(deepseek_v4): extract the eager attention reference into its own module
`eager_attention_with_sinks`, `build_sliding_window_mask` and
`block_bias_from_indices` all take plain tensors and none needs a
`PackedContext`, so they move to `eager_reference.py`, which imports nothing
from `attention.py`. Adds `dense_mask_from_indices`, which renders a gather
index tensor as the dense additive mask over a `kv_buf`, masking the trailing
sentinel position. That is the oracle the fused kernel is checked against once
`sparse_attention_gather` goes away: it exercises the index contract and the
attention math together rather than only the latter.

`attention.py` imports the three moved names as module globals, so callers and
the tests that monkeypatch them are unaffected. No behaviour change.
